requirements.txt
groq
fastapi
uvicorn
requests
python-dotenv
pydantic
PyGithub
streamlit
huggingface_hub
